// ts/types/takeout_schema.ts
export default interface RawTakeout {
	/**
	 * @minItems 1
	 */
	participants: [
		{
			name: string
			[k: string]: unknown
		},
		...{
			name: string
			[k: string]: unknown
		}[]
	]
	/**
	 * @minItems 1
	 */
	messages: [
		{
			sender_name: string
			timestamp_ms: number
			content?: string
			is_unsent?: boolean
			/**
			 * @minItems 1
			 */
			photos?: [
				{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				},
				...{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				}[]
			]
			/**
			 * @minItems 1
			 */
			gifs?: [
				{
					uri: string
					[k: string]: unknown
				},
				...{
					uri: string
					[k: string]: unknown
				}[]
			]
			/**
			 * @minItems 1
			 */
			reactions?: [
				{
					reaction: string
					actor: string
					[k: string]: unknown
				},
				...{
					reaction: string
					actor: string
					[k: string]: unknown
				}[]
			]
			share?: {
				link: string
				share_text: string
				[k: string]: unknown
			}
			/**
			 * @minItems 1
			 */
			videos?: [
				{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				},
				...{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				}[]
			]
			/**
			 * @minItems 1
			 */
			audio_files?: [
				{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				},
				...{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				}[]
			]
			/**
			 * @minItems 1
			 */
			files?: [
				{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				},
				...{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				}[]
			]
			sticker?: {
				uri: string
				ai_stickers: {
					[k: string]: unknown
				}[]
				[k: string]: unknown
			}
			ip?: string
			call_duration?: number
			missed?: boolean
			is_geoblocked_for_viewer: boolean
			[k: string]: unknown
		},
		...{
			sender_name: string
			timestamp_ms: number
			content?: string
			is_unsent?: boolean
			/**
			 * @minItems 1
			 */
			photos?: [
				{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				},
				...{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				}[]
			]
			/**
			 * @minItems 1
			 */
			gifs?: [
				{
					uri: string
					[k: string]: unknown
				},
				...{
					uri: string
					[k: string]: unknown
				}[]
			]
			/**
			 * @minItems 1
			 */
			reactions?: [
				{
					reaction: string
					actor: string
					[k: string]: unknown
				},
				...{
					reaction: string
					actor: string
					[k: string]: unknown
				}[]
			]
			share?: {
				link: string
				share_text: string
				[k: string]: unknown
			}
			/**
			 * @minItems 1
			 */
			videos?: [
				{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				},
				...{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				}[]
			]
			/**
			 * @minItems 1
			 */
			audio_files?: [
				{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				},
				...{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				}[]
			]
			/**
			 * @minItems 1
			 */
			files?: [
				{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				},
				...{
					uri: string
					creation_timestamp: number
					[k: string]: unknown
				}[]
			]
			type?: string
			sticker?: {
				uri: string
				ai_stickers: {
					[k: string]: unknown
				}[]
				[k: string]: unknown
			}
			ip?: string
			call_duration?: number
			missed?: boolean
			is_geoblocked_for_viewer: boolean
			[k: string]: unknown
		}[]
	]
	title: string
	is_still_participant: boolean
	thread_path: string
	magic_words: {
		[k: string]: unknown
	}[]
	image?: {
		uri: string
		creation_timestamp: number
		[k: string]: unknown
	}
	joinable_mode: {
		mode: number
		link: string
		[k: string]: unknown
	}
	[k: string]: unknown
}

export type RawTakeoutMessage = RawTakeout['messages'][0]